# api.yaml
openapi: 3.0.0
info:
  contact:
    name: Yannick Simon
    email: yannick.simon@kysoe.com
    url: https://yannicksimon.fr
  title: renovadmin v0
  description: prototype renovadmin v0
  version: 1.0.0
servers:
  - url: 'http://{{url_base}}'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    noauthAuth:
      type: http
      scheme: noauth
...
paths:
  /api/v1/company:
    post:
      description: |
        create company
      operationId: createCompany
      tags:
        - companies
      summary: create company kysoe
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties: {}
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    get:
      description: |
        list companies
      operationId: listCompanies
      tags:
        - companies
      summary: list companies
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /api/v1/user:
    post:
      description: create user
      operationId: createUser
      tags:
        - users
      summary: create user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                email: yannick.simon2@gmail.com
                password: '12345678'
                company: kysoe
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    get:
      description: list user
      operationId: listUsers
      tags:
        - users
      summary: list users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /api/v1/login:
    post:
      description: login process
      operationId: login
      
      tags:
        - General
      summary: login yannick
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                email: yannick.simon@gmail.com
                password: '12345678'
      parameters:
        - name: Content-ty
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /api/v1/token/refresh:
    post:
      description: refresh access token
      operationId: refreshAccessToken
      
      tags:
        - General
      summary: refresh token
      requestBody:
        content: {}
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}